Cleaned main, refactored seek help away from main file
